refactor: rename namespace to business rules engine
