+++ Src/Fitnet/Shared/BusinessRulesEngine/BusinessRuleValidationException.cs
@@ -1,4 +1,4 @@
-namespace SuperSimpleArchitecture.Fitnet.Shared.BusinessRule;
+namespace SuperSimpleArchitecture.Fitnet.Shared.BusinessRulesEngine;
 
 internal class BusinessRuleValidationException : InvalidOperationException
 {
