+++ Src/Fitnet/Shared/BusinessRulesEngine/IBusinessRule.cs
@@ -1,4 +1,4 @@
-namespace SuperSimpleArchitecture.Fitnet.Shared.BusinessRule;
+namespace SuperSimpleArchitecture.Fitnet.Shared.BusinessRulesEngine;
 
 internal interface IBusinessRule
 {
